Initial port from unittest to pytest
